+++ src/iris/gfx/frame_counter.cpp
@@ -5,14 +5,14 @@ namespace ir {
 
     master_frame_counter_t::~master_frame_counter_t() noexcept = default;
 
-    auto master_frame_counter_t::make(uint32 current) noexcept -> arc_ptr<self> {
+    auto master_frame_counter_t::make(uint64 current) noexcept -> arc_ptr<self> {
         IR_PROFILE_SCOPED();
         auto master = arc_ptr<self>(new self());
         master->_current_frame = current;
         return master;
     }
 
-    auto master_frame_counter_t::current() const noexcept -> uint32 {
+    auto master_frame_counter_t::current() const noexcept -> uint64 {
         IR_PROFILE_SCOPED();
         return _current_frame.load(std::memory_order_acquire);
     }
@@ -53,12 +53,12 @@ namespace ir {
         return *_master;
     }
 
-    auto frame_counter_t::current_absolute() const noexcept -> uint32 {
+    auto frame_counter_t::current_absolute() const noexcept -> uint64 {
         IR_PROFILE_SCOPED();
         return master().current();
     }
 
-    auto frame_counter_t::current_relative() const noexcept -> uint32 {
+    auto frame_counter_t::current_relative() const noexcept -> uint64 {
         IR_PROFILE_SCOPED();
         return current_absolute() - _current_frame;
     }
